feat: extract exchange handling to separate module
